Fixing JDT errors that java compiler doesnt catch but JDT catches : )
Fixing JDT errors that java compiler doesnt catch but JDT catches : )

+++ src/main/java/site/repository/RegistrantNumberGeneratorRepository.java
@@ -7,7 +7,7 @@ import site.model.Registrant;
 
 @NoRepositoryBean
 public interface RegistrantNumberGeneratorRepository<T extends Registrant.NumberGenerator> extends
-    JpaRepository<T, Long> {
+    JpaRepository<T, Integer> {
     T findFirstByOrderByIdAsc();
 
     default Class<?> repositoryClass() {
